Add order module
Add an order sub-module to the orderbook module.

This patch makes no logic changes.
