+++ comit/src/network/protocols/orderbook.rs
@@ -1,7 +1,7 @@
+mod order;
 mod take_order;
 
 use crate::{
-    asset, ledger,
     network::protocols::orderbook::take_order::{Response, TakeOrderCodec, TakeOrderProtocol},
     SharedSwapId,
 };
@@ -26,7 +26,8 @@ use std::{
     task::{Context, Poll},
     time::Duration,
 };
-use uuid::Uuid;
+
+pub use self::order::*;
 
 // We only support a single topic at the moment.
 const TOPIC: &str = "Herc20Hbit";
@@ -266,8 +267,6 @@ impl TradingPairTopic {
     }
 }
 
-pub type OrderId = Uuid;
-
 /// MakerId is a PeerId wrapper so we control serialization/deserialization.
 #[derive(Debug, Clone, PartialEq)]
 pub struct MakerId(PeerId);
@@ -309,60 +308,6 @@ impl<'de> Deserialize<'de> for MakerId {
         let peer_id = PeerId::from_str(&string).map_err(D::Error::custom)?;
 
         Ok(MakerId(peer_id))
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-pub struct Order {
-    pub id: OrderId,
-    pub maker: MakerId,
-    pub buy: u64,
-    pub bitcoin_ledger: ledger::Bitcoin,
-    pub sell: asset::Erc20,
-    pub ethereum_ledger: ledger::Ethereum,
-    pub absolute_expiry: u32,
-}
-
-#[derive(Debug)]
-pub struct NewOrder {
-    pub buy: asset::Bitcoin,
-    pub bitcoin_ledger: ledger::Bitcoin,
-    pub sell: asset::Erc20,
-    pub ethereum_ledger: ledger::Ethereum,
-    pub absolute_expiry: u32,
-}
-
-impl Order {
-    pub fn new(peer_id: PeerId, new_order: NewOrder) -> Self {
-        Order {
-            id: Uuid::new_v4(),
-            maker: MakerId(peer_id),
-            buy: new_order.buy.as_sat(),
-            bitcoin_ledger: new_order.bitcoin_ledger,
-            sell: new_order.sell,
-            ethereum_ledger: new_order.ethereum_ledger,
-            absolute_expiry: new_order.absolute_expiry,
-        }
-    }
-
-    pub fn topic(&self, peer: &PeerId) -> Topic {
-        TradingPair {
-            buy: SwapType::Hbit,
-            sell: SwapType::Herc20,
-        }
-        .to_topic(peer)
-    }
-}
-
-impl NewOrder {
-    pub fn assert_valid_ledger_pair(&self) -> anyhow::Result<()> {
-        let a = self.bitcoin_ledger;
-        let b = self.ethereum_ledger;
-
-        if ledger::is_valid_ledger_pair(a, b) {
-            return Ok(());
-        }
-        Err(anyhow::anyhow!("invalid ledger pair {}/{}", a, b))
     }
 }
 
@@ -521,6 +466,7 @@ mod tests {
     use super::*;
     use crate::{
         asset::{self, Erc20, Erc20Quantity},
+        ledger,
         network::test::{await_events_or_timeout, new_connected_swarm_pair},
     };
     use libp2p::Swarm;
